试运行通过

+++ src/main/java/com/datech/zjfh/api/service/BizAlarmConfigTask.java
@@ -12,7 +12,8 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 
 import javax.annotation.Resource;
-import java.util.Calendar;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.List;
 
 @Slf4j
@@ -27,10 +28,10 @@ public class BizAlarmConfigTask {
     @Resource
     private BizIvsServiceImpl bizIvsService;
 
-    @Scheduled(cron = "1 * * * * ?")
+    @Scheduled(cron = "0 * * * * ?")
     public void configureTasks() {
-
-        String nowTime = getNowTime();
+        SimpleDateFormat format = new SimpleDateFormat("HH:mm:ss");
+        String nowTime = format.format(new Date());
         log.info("====BizAlarmConfigTask, nowTime:{}", nowTime);
         LambdaQueryWrapper<BizAlarmConfigEntity> configQuery = Wrappers.lambdaQuery();
         configQuery.eq(BizAlarmConfigEntity::getState, 0);
@@ -39,46 +40,37 @@ public class BizAlarmConfigTask {
             return;
         }
         for (BizAlarmConfigEntity config : configList) {
-
             if (nowTime.compareTo(config.getBeginTime()) >= 0 && nowTime.compareTo(config.getEndTime()) < 0) {
                 //在非布防时间内
                 BizCameraEntity camera = bizCameraService.getById(config.getCameraId());
-                if (StringUtils.isNotBlank(camera.getSubscribeId())) {
+                if (camera != null && StringUtils.isNotBlank(camera.getSubscribeId())) {
                     int resultCode = bizCameraService.deleteIntelligentData(camera.getSubscribeId(), bizIvsService.getById(camera.getIvsId()));
                     if (resultCode == 0) {
                         camera.setSubscribeId("");
                         camera.setSubsEnable(0);
                         bizCameraService.updateById(camera);
-                        log.info("摄像头:['" + camera.getId() + "']删除订阅告警，时间:" + nowTime +"， 非布防时间段："+config.getBeginTime()+"--"+config.getEndTime());
+                        log.info("摄像头:['" + camera.getDeviceIp() + "']删除订阅告警，时间:" + nowTime +"， 非布防时间段："+config.getBeginTime()+"--"+config.getEndTime());
+                    } else {
+                        log.info("摄像头:['" + camera.getDeviceIp() + "']删除订阅告警失败！！！！！，resultCode:" + resultCode);
                     }
                 }
             }
-            if (nowTime.compareTo(config.getEndTime()) == 0) {
-                //当前时间等于非布防时间段结束时间，订阅告警
+            if (nowTime.compareTo(config.getEndTime()) >= 0) {
+                //当前时间大于等于非布防时间段结束时间，订阅告警
                 BizCameraEntity camera = bizCameraService.getById(config.getCameraId());
-                if (camera != null) {
+                if (camera != null && StringUtils.isBlank(camera.getSubscribeId())) {
                     //订阅告警
                     int resultCode = bizCameraService.addIntelligentData(camera, bizIvsService.getById(camera.getIvsId()));
                     if (resultCode == 0) {
                         camera.setSubsEnable(1);
                         bizCameraService.updateById(camera);
-                        log.info("摄像头:['" + camera.getId() + "']新增订阅告警，时间:" + nowTime +"， config："+config.getBeginTime()+"--"+config.getEndTime());
+                        log.info("摄像头:['" + camera.getDeviceIp() + "']新增订阅告警，时间:" + nowTime + "， 非布防时间段：" + config.getBeginTime() + "--" + config.getEndTime());
+                    } else {
+                        log.info("摄像头:['" + camera.getDeviceIp() + "']新增订阅告警失败！！！！！，resultCode:" + resultCode);
                     }
                 }
             }
         }
     }
 
-    private String getNowTime() {
-        StringBuilder nowTime = new StringBuilder();
-        Calendar today = Calendar.getInstance();
-        nowTime.append(today.get(Calendar.HOUR_OF_DAY));
-        nowTime.append(":");
-        if (today.get(Calendar.MINUTE) < 10) {
-            nowTime.append("0");
-        }
-        nowTime.append(today.get(Calendar.MINUTE));
-//        log.info("nowTime:{}", nowTime);
-        return nowTime.toString();
-    }
 }
